fix(sessions): satisfy durable snapshot review gates
